HeaderTitle: restructure as dir + cleanup
